Splits Dashboard into components
